+++ lib/commands.js
@@ -9,10 +9,23 @@ try {
 	commandDirectory = '/modules';
 	commandFiles = Kirbi.getFileArray(commandDirectory);
 } catch (err) { 
-	Kirbi.error(err);
+	Kirbi.logError(err);
 }
 
 exports.init = function() {
+	// set up helpers
+	Kirbi.Commands = {};
+	Kirbi.addCommand = function (commandName, commandObject) {
+		try {
+			Kirbi.Commands[commandName] = commandObject;
+		} catch (err) {
+			console.log(err);
+		}
+	}
+	Kirbi.commandCount = function () {
+		return Object.keys(Kirbi.Commands).length;
+	}
+
 	// Load command files
 	commandFiles.forEach(commandFile => {
 		try {
@@ -59,10 +72,9 @@ exports.init = function() {
 	if (Kirbi.Config.externalModules && Kirbi.Config.externalModules instanceof Array) {
 		Kirbi.Config.externalModules.forEach(module => {
 			if (Kirbi.Commands[module]) {return};
-			//grab from node_modules
-			var dir = '/node_modules/kirbi-' + module;
+			var moduleName = 'kirbi-' + module;
 			try {
-				module = Kirbi.require(`${path.join(dir, module)}`);
+				module = Kirbi.require(`${module}`);
 			} catch (err) {
 				Kirbi.logError(`Improper setup of the '${module}' command file. : ${err}`);
 				return;
